Change ghost state cache representation
This is a large portion of a major change from representing the cache
as an AssocCache (a Map.t cache_entry) to a total function, including
(Invalid, NoOwner) entries for the uncached addresses.

While this makes the existing proofs hard to re-use, the new
representation and functions on it should be simpler enough to make the
result better and more sustainable.