Model locked frame in a more composable way
- Relatively simple definition: lift a predicate to also say its
addresses are locked. We also mention that the predicate is precise, but
don't currently use it; without it, however, the lockset is ambiguous
from the predicate.
- Simple example of absorbing an address into the locked frame; moving
the (a |-> v) pred is straightforward since it is not only precise but
strictly_exact and has an obvious domain (a).
- File is now very messy, with multiple ways of accomplishing
essentially the same thing; F * locks_held LF should supersede
split_frames F LF.