Prove dwrite_ok in Log.v
Spec for dwrite_ok(a, v) turns out to be quite interesting:

    PRE
      rep xp F (ActiveTxn m1 m2) ms * [[ a <> 0 ]] *
      [[[ m1 ::: (Fm1 * a |-> vs1) ]]] *
      [[[ m2 ::: (Fm2 * a |-> vs2) ]]]
    POST RET:ms' exists m1' m2',
      rep xp F (ActiveTxn m1' m2') ms' *
      [[[ m1' ::: (Fm1 * a |-> (v, vsmerge vs1)) ]]] *
      [[[ m2' ::: (Fm2 * a |-> (v, vsmerge vs1)) ]]]

Notice that in the post condition, both starting and ending disks
(m1, m2) evolve to new disks (m1', m2').  This is different from
write_ok, where the starting disk m1 remains intact in the
post-conditon.

If a transaction only invokes write(), we will know that m1 remains
the same during the entire transaction as before. Thus log_recover
can roll back to the state when begin() is called.  There is no
extra prove overhead at all!

However, if dwrite() is ever called, we will get a different starting
disk state m1' thereafter. Only a frame Fm1 is preserved across
m1 and m1'.  Higher-level code is responsible for its own recovery
using this information.

Also note that vs2 is not used in the post condition, because
dwrite() will discard any pending updates to the same address in the
transaction's memstate.