BFile.v: an attempt for dwrite_ok spec
This spec is very verbose, it describes the state of "abort disk"
and the "commit disk" in both pre and post conditions.

The statements are exactly alike for both disks, but we have to use two
sets of variables for frames and objects from either disks.

Due to too many variables, the automation becomes extremely slow.
It's getting stuck in the "destruct_product" step when normalizing.
So I cannot keep going.

Actually the spec is talking about two parallel worlds (the abort
world and the commit world) in a mingled way.  For a given operation
(like dwrite), its effects on one world just mirror the other.  Ideally
we should only reason about one of them, and automatically derive the
state of the other.  Not sure what's the best way to do this.