GroupLog.v: try deferred flush
GLog is a new layer in between Log and MemLog.  It defers flushing
of log entries when a transaction commits.  Instead, all committed
txns are cached in the memory.  When flushall() is called,
GLog flushes each one in order, by calling MLog.flush().

The abstraction GLog exposes is a sequences of virtual disks.
The disk sequence satisfies prefix property: Each disk represents
a checkpoint when a transaction commits, and is derived from its
predecessor.  The first one corresponds to the last transaction
that has been flushed to the disk; and the last disk maps to the
most-recently committed transaction, which is the disk that the higher
layer is supposed to see.

If flushall() crash in the middle, the recover code can bring the disk
back to any of those states, so that any transactions happens before it
is also persisted.