half of a writeback cache
read and sync are implemented and proven correct

write is almost doable; the only tricky part seems to be, what to
do about multiple writes to the writeback-cache?  these multiple
writes should stack up in the valuset, just like what happens when
we write to disk directly.  but since the writeback-cache does not
keep around old writes, it's not clear where they should be coming
from in the spec.

thanks to Adam's suggestions and Bedrock's [starB] for inspiring
MemPred.v, which made WritebackCache.v far simpler (and should
probably be used to make Cache.v simpler too).