Checkpoint rewrite to do fix specs and cleanup proofs.
The specs for describing DstFileExists where too lose: the user of the spec
couldn't conclude that the only thing that changed in the directory tree was the
creation of the destination file.

Many dirtree lemmas are unproven, but maybe let it be, because we are switching
to Haogang's dirtrees.

Todo: eliminate a few easy admits and update FS.delete spec (maybe that also
should be done in the disklog2 branch).