Define linear_rel
- Proved it maintains consistency under the lock protocol. The
definition is fairly complicated; it would be easier to specify what is
allowed to change, but instead it describes what addresses must be
equal. The problem is largely that a linear_mem has too many degrees of
freedom: every address is really just one value (if unlocked) or two
values (if locked).