Implement cache lock
Discovered something wrong about how preserves works:

preserves R F F' says that forall P,
 s |= F * P ->
 R s s' ->
 s |= F' * P.

This is too strong: P should be allowed to evolve according to the
relation. The point of preserves was really to factor R, but it seems
it needs a factoring specific to both F and the starred P, which seems
really hard to prove.