prove a stricter spec for file_getattr_recover_ok
For read-only transactions, we know that in REC, the disk
must be the crash_xform of the *first* disk in the diskset.
This is stricter than after_crash, which says the disk
can be the crash_xform of any of the disks.

Not sure if this is useful/necessary or not, because PRE
says nothing about the first disk anyway.