Prove create_recover_ok and rename_recover_ok
Two proves are identical. The key part is to instantiate
idemcrash to include the after-crash case.

As we are always going back to the first disk, just write

    [[[ d ::: crash_xform (diskIs (list2nmem (fst ds))) ]]]

in REC. No need for exists n, nthd ...

Because we never chain REC specs together, the REC condition
is actually the end of the story (no caller would ever use
it as the precondition). We can just leave the crash_xform,
and do not worry about crash_xform of Tree.rep

But leaving crash_xfrom (diskIs ...) makes the spec less readable.
It is still desirable to describe in REC what's the state of
the tree after recovery, using the proposed tree_crash predicate.

For recover_ok in App.v, neverthelss, we definitely need tree_crash
to help us reason about the tree's state after crash.