Finish locked_AsyncRead_ok with some cleanup
Factoring out the stability proof into an alternate spec for Yield
simplifies things - the right structure is probably to have that spec
be local to the module, with the _ok prog hint unexported, so it isn't
used elsewhere, since it is specific to what is stable for this module.

More generally though it might be useful to allow instantiating Yield_ok
with a stable fact to apply to the postcondition (or automating this
pattern for the many specs that will have star R' in their
postcondition).