tweak the WritebackCache rep so that the evict_ok spec is simpler
i think now you should be able to prove that [evict] is a noop (does
not change the WritebackCache rep).  but this slightly breaks all the
WritebackCache proofs.  so i'm not sure if it's an end-to-end improvement
(might still be a good idea, since then we don't have to change the
proofs in higher-level functions that call evict and friends.)