Make progress on get_item_ok
Dealt with readers by having specs specifically say nobody is reading
the particular block. This is an "invariant" in the sense of being in
the pre/post of every method, but from the system's perspective it
can't be in Inv since the LockedDisk will have to yield after starting
a read.

We would like to say nobody is reading any RecArray address (which is a
simpler invariant), but we can only guarantee this for locked items.

Much of this proof is list manipulation. Proved some of these already,
so they do not represent most of the remaining work.