state a better spec for AtomicCp's [recover]
there was a tricky issue with [exists] vs [forall] in the postcondition.
when we refer to representation invariants that should have already been
true of prior disks, using [exists] the way we did:

  exists tree,
  [[[ some_disk_from_old_ds ::: DIRTREE.rep tree ]]]

is a bad idea, because that effectively states that there's a possibly
different representation invariant for the same disk!  several steps
later, in the corr3 proof, we get stuck having to prove that, because we
have two [DIRTREE.rep] facts about the same disk, the trees must be equal!
but that's just not true; there's many equivalent trees that are not
strictly equal to each other.

let's try using [forall] instead: the postcondition of [recover] says
that, if someone can prove that a disk from the original diskset satisfies
a [DIRTREE.rep], then something nice holds true in the resulting disk's
tree as well.

we'll probably still need to add something about [DIRTREE.rep] in the
precondition, just so that we can safely run tree operations inside the
recovery code..