Try to use the facts about dirtree transforms to prove a more end-to-end fact
This turns out to be tricky, due to annoying corner cases.  Namely, suppose we
are modifying some block that belongs to an inode in the current disk.  These
two cases are hard to prove right now:

(1.) What if, in an older disk, the same inode number corresponds not
to a file but rather to a directory?

(2.) What if, in an older disk, the block belongs to the same inode
number, but that inode isn't in the tree?

Need to figure out what invariant would help us prove these cases..