a better version of the dirtree safety invariant, [dirtree_safe2]
instead of preventing inode reuse, which is messy, i think it suffices
to say that (1) blocks will not be reused for a directory with the same
inode number -- i.e., it must be a file, and (2) if a block is not found
in a file inode with the same inode number, it must be on the freelist.

the first is just a proof obligation.  the second requires changing our
code slightly: instead of calling BFILE.reset when we first allocate a
file, we should call BFILE.reset when we delete a file (e.g., when
[delete] or [rename] calls [SDIR.unlink]).