No need to call flushall() in [dsync] / [dsync_vecs]
Even if there are pending transactions writing to an address,
we are still safe to dsync that address. (most likely when we
first call dwrite() on that address, flushall() will be called).
This change further simplifies top-level file_sync_ok spec,
as there is no branches in its post condition.

In terms of spec and proof, the dssync/dssync_vecs operators are
commutative with replay_disk -- this is not true for dsupd, so
we might need to call flushall in dwrite().