run-bench.sh: explicitly say data=journal in async_journal_commit mode
interestingly, ext4 will happily mount a file system with BOTH async and ordered mode,
if you specify the async_journal_commit flag, but omit any specification of data=.
under the covers, it will then default to data=ordered, even though if you specify
this explicitly in the mount options, it will refuse to mount..

we're not better than linux in terms of syncs (same number of syncs to journal, because
linux is presumably capable of doing group apply).  we need to figure out how to get
linux to flush its journal (presumably run with a sufficiently large file so that it
overflows the journal and has to apply), and also we need to measure the number of
writes to show why we win.