Messy progress toward cache_commit
No need to do with a loop, since all operations are in memory.
Modelling and reasoning about emptying the cache is surprisingly
tricky, though, and needs a lot of NoDup list manipulation. The
standard library provides very few theorems in this area, so many basic
lemmas are proven here and ought to be cleaned up eventually.