[crash_xform_rep_r] is no longer true given [ptsto_subset]
We want to show that

    possible_crash m m' ->
    rep cs' m' =p=> exists cs, crash_xform (rep cs m).

The right-hand-side of pimpl requires the underlying memory to be synced
(all tails are nils).  However, even though from [possible_crash] we know
that m' is properly synced, [rep cs' m'] might not describe a synced memory.

The problem is that [rep] uses [ptsto_subset]. Unfortunately,
"incl tail (vsmerge (v, nil))" is too weak: we want to prove the tail is
nil, but in fact tail can by any repetition of [v].