Fix indirect rep (hopefully for the last time)
This rep is a simplified version of the previous rep that eliminates the
listmatch over firstn terms. The listmatch is now over entire lists, which
should make proving much easier. It also specifies conditions for the entire
list of indirect blocks, substituting 0's for addresses at indices greater than
the valid pointer count. Specifying the entire list, as opposed to the first
nvalid entries makes it easier to reason about while not imposing any additional
restrictions on the disk, as all these 0 pointers are unused. The changes are
based on a rep suggested by Nickolai, but with additional tracking of the number
of valid blocks throughout the tree, as opposed to just on the bottom layer.