Checkpoint progress on indshrink
Shrinking is hard, so instead of having a single function that does it, shrink
will be broken up into 3 functions: the first will implement shrinking for a
tree where both the starting number of valid blocks and the final number of
valid blocks are (NIndirect ^ indlvl)-aligned, so that each subtree referenced
by the top-level tree is either fully valid or fully unused. This function
can be implemented recursively in a trivial manner. The second function will
reduce shrinking a tree where one of the top-level children is only partially
valid to the aligned case. This will be implemented recursively, and will also
make use of the aligned shrink. The last function will handle the case where
the highest-index valid child has an aligned valid count and the new valid
count is unaligned but that child is still partially valid.

This commit adds the first function, which handles aligned shrinks, and proves
it for indlvl = 1. The proof is a mess, but will be cleaned up in a later
commit.