Fix cache using relaxed FinishRead semantics
Readers no longer represent a lock, since other threads could finish the
read. On the other hand, there's no real need for this: we ensure we
only finish reads that somebody started by using [cache_rep] and
checking for [Invalid] in the cache.

This should have been true from the beginning, but now yields aren't
even in the cache code.