Simplify indirect block rep and re-prove functions
Instead of keeping track of the number of valid blocks at each level of the
tree, and avoiding specifying anything about the invalid blocks, this rep
specifies the full tree, specifying that any 0 entries "point" to a virtual
all-zero block. This makes the functions much easier to prove, as eliminates
the need to change the predicate mapped by listmatch, which was painful under
the previous scheme. Since the tree is fully specified at all times, this should
make it fairly easy to support sparse files at the block level in the future.
The eventual goal is to specify and prove the following operations
on indirect blocks:
  - get  : get the block number at a specified index in the linear block list
  - read : read the entire linear block list
  - put  : set the block number at a specified index to a nonzero value
  - clear: set the block number at a specified index to 0 - this should
           also reclaim unneeded indirect blocks.

This patch specifies and proves the first three operations, though the proof for
`put` is quite messy and will be cleaned up in another commit.