Progress proving cache simulates Prog.prog
The proof has some trickiness due to how failures are handled. The
cache_read/cache_write specs promise not to fail, but simulation is
stated in a way that failures must be identical.

It's possible this can be sidestepped by changing what the theorems say
about failure. On the other hand, it's not the case that failures are
currently correct simulated - cache_write can succeed even if the
address is out of bounds, whereas the Write will fail. On the other
hand, such a program would break the cache invariant and would never be
able to call Yield again.