Add for loop to Go semantics
The Go.For loop takes a variable, termination expression, and body, and
executes the body as long as the variable is less than the termination
expression. After each execution of the body, the variable is
incremented by one. This corresponds to a very limited for loop - it is
over one variable, does not initialize the variable, and increments
exactly by one every time. The reasoning here is that the body and
termination condition are fairly general, so without restricting the
body to progs that don't modify the variable, it's not reasonable to
bind the variable to a value for each iteration of the loop, which was
another possible incarnation.

With this patch, the Go.For loop  exists and is well-defined, but there
is no extraction code to produce it yet.