Prove exec covers every program * state
Specifically, every program can execute to completion or failure, or we
can derive a hash collision from a program's use of hashing.

Hash collisions are stated in terms of a natural definition, which has
to be related to hash_safe, which is the non-collision case given in
the semantics. In addition, these properties only hold when the hashmap
respects an invariant corresponding to only including valid hashes; we
prove this well-formedness condition is preserved by execution, so the
final theorem only requires starting with a well-formed hashmap (and in
particular an empty hashmap is trivially well-formed).