One tree_inodes_distinct admit left; almost done with treeseq?
Needed to update AFS.rename_ok to include:
 [[ forall inum def', inum <> dnum ->
              selN ilist inum def' = selN ilist' inum def' ]]
but that might be too strong---will fscq also change the inode of the file being
deleted?

Note that if we prove the last admit in treeseq, then we still need to fix
the proofs for AFS.rename_ok and AFS.delete_ok, since we changes their specs.