Use CompileConst only on constants
The compile_const tactic was matching non-constant expressions and
using CompileConst to solve them. Since parts of those expressions are
not in the scope of the Go statement being compiled, this is illegal.
Coq, however, fails to recognize this until the Define at the end, at
which point it fails. This adds a check to force compile_const to only
compile constant expressions.