Add a postcondition to read_fblock_ok
The automation seems to be pretty brittle, will need to fix something.
One issue in particular is that the [get _ s] values tend to not exactly
match due to different implicits - there's an id somewhere that should
be (fun T => T) so it never has to be unfolded to compute.