add some machinery for knowing when some procedure is read-only with respect to the memory state
this might be useful for parallelizing read-only operations in the future

currently this is exposed as [[ readOnly ms ms' ]] at the Log.v API; need
to propagate this further up the stack if this turns out to be useful.

might be useful to hide the conditional [exists] so the statement doesn't
look too awkward in higher layers.