pass in [auto with *] as the default [intuition_t] tactic
Omega re-defines auto as [auto with *] using Ltac.

This is dangerous because [auto] is defined in Coq's implementation
as notation for the [auto] tactic.

Somewhat surprisingly, depending on how [auto] is passed as an argument,
it sometimes appears to get the built-in [auto] through the notation,
and other times it appears to pick up the Omega-defined [auto] Ltac!
For instance, in v8.5, [intuition auto] picks up the built-in [auto],
but this code picks up the Omega-defined [auto with *]:

  Require Import Omega.
  Lemma z : forall i, i < 0 -> False.
  Ltac intuition2 myauto := intuition myauto.
  intuition2 auto.

In Coq-trunk this behavior seems to have changed, so that more often
we got the built-in [auto].  This was less powerful, and caused some
proofs to break.

This change passes [auto with *] explicitly to [cancel_with'] to try
to get the behavior we used to have in v8.5.