Make [writeback] compile with coq 8.6
Explicitly declaring the variables caused some of the type parameters to
fst and snd to be unfolded. This prevented cancel from matching terms
correctly, thus preventing compilation. Now use matches to grab the
literal terms from the goal for correct cancellation. Also, the "Time
Define" doesn't work with Coq from the command-line, so remove the
"Time".