Fix compile_map_op and compile const booleans
Make compile_map_op perform the entire compilation step instead of
aborting halfway with a TODO message.

Also compile (Ret true) and (Ret false) automatically. Since we
currently only support numerical literals, these are compiled as the
result of always true expressions: `Ret true` is compiled as the result
of the expression "a == a" where a is some arbitrary variable;
`Ret false` is treated similarly.