Speed up cancel_go_fast
Some work was being duplicated when matching ptstos on the right and
left. Since the search tactic that matches up ptstos already traces the
entire sep_star tree on the left, preserve the path and use it to
move the correct ptsto all the way to the left for matching. This
reduces the "Defined" time for compile_writeback to 25 seconds.