Change concrete Go extraction pipeline
This commit attempts to solve several problems at once:

1. compiling extraction theorems is expensive
2. extraction is only necessary for building a binary, not for proving
3. each file with progs has an independent subset of the FSCQ code
4. proven extraction requires these dependencies to be explicitly stated

The first problem suggests separating the _ok proofs about code from the
extraction proofs, since the latter are run to completion even for .vio
files. To this end, the extraction code is in the src/Extract directory.
make's dependency resolution should ensure that these are only built
after a successful build of all the other .vio files, which mostly
solves the second problem.

The solution to problems 3 and 4 is a little messier. In order to keep
the makefile as the single source of truth for dependencies, each
Extract/ExtractXXX.v will declare a compilation theorem for each of the
functions from XXX.v that needs to be extracted. These will all be
nicely bundled in a function/theorem that takes as input facts about all
the function calls to code outside XXX.v, and that produces as output an
Env map of all the compiled functions in XXX.v. By composing these
functions in the right order (which is always possible - just flatten the
dependency DAG), an environment for the entire FSCQ code can be
constructed.

This approach has two main drawbacks: the code that composes these maps
must either take an unordered list and resolve dependencies or have the
maps listed in a correct order. Since the dependencies are unlikely to
change often, this is not too bad. The other downside is that all
extracted code will be spit out as one huge Go file. This isn't great,
but it's the plan for now.