Fix function call compilation
The SetVarBefore theorem was too restrictive; it required that the
value being set made no modifications to the rest of the state, which is
impossible if the value being set performed any destructive operations.
This now ensures that function arguments are compiled in order. In
compile_call, each argument that does not yet exist is produced; if all
exist, the call is extracted. As a small hack, each argument must exist
twice, since the function call will consume one copy. This is a strategy
that might be useful elsewhere as well to prevent loss of values.