interesting issue with extracting Balloc
[IAlloc.Alloc.Bmp.ifind] takes a callback function as an argument --
this function decides if a given bit position matches what the caller
wanted, or whether [ifind] should keep looking for other bit positions.

We currently have no wrappers for functions.

What's the right plan here and in general?

Plan 1: change this particular [ifind] to take a flag (to indicate which
of two possible callbacks it should use), instead of the literal callback.
But we might run into this issue again in [Dir] where we use [ifind]
in a more sophisticated fashion.

Plan 2: add support for function pointers to Go semantics.  This seems
like it could easily spiral out of control.

Plan 3: specialize extraction of [ifind] with specific callback arguments,
perhaps pragmatically re-ordering that callback argument to be the first
argument to [ifind], so partial applications are easy.  This seems like
it might work.

What does Facade do?