Make DeepCopy to take dest ptr, make buffers lazy
Since all variables are initialized to something reasonable, reuse that
memory in DeepCopy() instead of ignoring it and returning a new object.

Also, since many Buffers are allocated, initialized, and then
overwritten, don't allocate a large slice internally until the space is
actually used. This does make things more complicated on the Go side but
brings the cache-heavy read workload performance to within a factor of 2
of Haskell's.