Start translator with simulation proof
The concurrent code has fewer error cases than the sequential, due to
CacheWrite never erroring (though it does break the cache rep on an
out-of-bounds write). This should be safe for transferring specs, since
we could conservatively have OOB writes fail and the sequential spec
would prove this doesn't happen.