Finish translator proof by allowing seq failure
It is safe to allow for any execution, rather than proving simulation at
the level of constraining the concurrent output, showing that the
sequential program fails. Verified sequential programs don't exhibit
this behavior, and the error cases in the sequential semantics
correspond to invalid (though possible safe) uses of the cache.