Prove equivalence between two Hoare triples
One is the natural definition, directly in terms of execution, the other
is about arbitrary continuations and uses a Hoare double. This is a
clean statement that the reflected triples are given exactly the right
meaning.

It seems like the Hoare double definition is generally more useful
(specs chain nicely), but for any metatheory the triples are easier to
reason about.