Translate a TreeSeq spec to syntax using notation
That is, slightly modify the existing spec notation to parse as a Hoare
double using a SeqSpec record for the spec. This spec can be
automatically translated in [translate_ok]. Specs require one small
modification: the line of variables after the "{<" have to be turned
into a single pattern binding, turning [ds ts pathname] into ['(ds, ts,
pathname)], for example.

Currently the spec is actually expressed as A -> rawpred -> SeqSpec; the
A is the existentially quantified tuple to replace the ghost state in
the spec notation, the rawpred is the frame added by the spec notation.
These specs should still be convertible, moving those two parameters
into the concurrent Spec's ghost variable (where there's currently a
unit).