Fix handling of frame predicates
Solution is quite simple: most definitions are completely general,
except for the notation, which produces a function rawpred -> SeqSpec A;
this can be turned into an appropriate SeqSpec (rawpred * A) by
inserting the uncurrying. Now the conversion only needs to be applied
when parsing the notation, and other code is unaffected.