Switch to separation logic over memory
Unify ghost state and memory into a single memory from variables to
arbitrarily-typed values (distinguishing between ghost and regular
variables). Variables are now dynamic, with a primitive to allocate
them. Specs will use separation logic over memory variables, with
assertions capturing that uses of variables to Gets and Assgns are
well-typed (Get returns a value of the expected type, failing otherwise;
the precondition guarantees this doesn't happen, and Assgns are similar,
guaranteeing the type of a variable doesn't change).