State correctness of readonly_syscall
Highlights a weakness of the translator specs: need to track that under
a read lock, the postcondition can be very strong and state that nothing
changed. Should essentially have separate specs (even if they're
factored into a single spec with two postconditions, one for ReadLock
and the other for WriteLock).