Make Balloc use a two-level bitmap allocator
Instead of using RecArray to serialize a list of bits, use it to
serialize a list of words, then deserialize only a single word into a
list of bits.

This does waste some additional space over the bit-based version,
because instead of refusing to allocate the first bit, it refuses to
allocate any bits in the first word. This might be fixed later.