Coordinate only for necessary memory writes
First, implement an RCU-like plan for the memory where system calls read
it from an IORef and update it only with a write lock. This avoids
acquiring a read lock, which is itself expensive in terms of
coordination with concurrent clients.

Second, we still need to update the memory only initially when the cache
is being constructed. To do so, update the memory based on a global
flag, and set it after warmup (the implementation toggles this flag on
getFileSystemStat, the statvfs handler, and fsbench runs df after
warmup).