Remove admit in reflective cancelling
Instead of continuing to try to prove the version of cancel_some that
removes the first matching right entry for an entry (which is annoyingly
difficult), exploit the fact that the lists passed to cancel_right and
cancel_left are sorted, and should have cancellable elements that line
up. Then cancel_some only needs to try to cancel elements at the same
indices.