Make cache purely thread-local
The cache now has no memory state, buffering purely in a state monad
with a cache (and no write buffer). Writes and proper cache misses (ie,
those that fill the cache) still require a write lock to protect the
disk. Non-filling reads are safe for any disk, whereas other operations
require that the cache rep apply to the physical disk, which can be
guaranteed with a write lock and by starting from a known good cache
(eg, by reading from memory and using the invariant).

Without a write buffer, there is no need for CacheCommit and CacheAbort
- higher levels will have to worry about writing the cache to memory,
and can abort by reverting to an earlier cache and its corresponding vd.