Fix up file system implementation
Proofs work only up to translator, with file system proofs not yet
started. Specifically, we need a generic proof of read-only system calls
in terms of a specification for the translated, optimistic system call,
and then to make sure that the general form for optimistic specs is
correct we need to prove a system call (OptFS.file_get_attr) meets its
reflected spec.

More interesting is to prove that the write syscall wrapper works, which
has both a loop and requires proving that the write at the end respects
the protocol.