Add construct to Prog to signal changes in caches
The new [AlertModified] constructor to Prog is isomorphic to [Ret tt]
(and thus doesn't do anything), but we can use it from the concurrent
interpreter to know when the new memory state is important and should be
incorporated.

This commit adds calls to AlertModified in the miss case for the buffer
cache and the directory cache. It's not important to have the flag be
accurate for modifying system calls, since those will require writing
the new state anyway. However, there may be another cache somewhere I
missed.