Finish opt_file_get_attr_ok
This commit combines two slightly different problems:

First, the AsyncFS spec file_getattr_ok was too weak: the DirTree.rep
now depends on mscs, so it is necessary to re-prove the rep even though
the disk hasn't changed. This is handled by the automation but requires
the additional postcondition to be added through a few layers. This
commit only fixes file_getattr_ok; other read-only system calls need the
same fixes.

Second, there were some automation issues related to hashmaps that made
the opt_file_get_attr_ok proof not as short as it could be. The proof
has been simplified from a hacky first attempt. A similar proof is
needed for each system call; all of it should be factored out into a
single tactic to the extent possible.