Allow OptimiticFS to be compiled wwout proofs
Previously took 130s to compile; now takes 2.7s.

Each theorem had to be proven to determine whether it depended on the
section variable G (otherwise the external signature was not known - it
may or may not take G as a parameter). This was required even under coqc
-quick. The fix is to explicitly depend on G with [Proof using] (in the
form of a [Set Default Proof Using]), which lets the STM know to skip
all the _ok proofs.