Preserve lock state across rely
Lock state really shouldn't be part of the protocol, since it's a local
view of the actual global lock state (which for a write lock is just
either unlocked or an owning thread id). The local lock state is
unaffected by other threads - you never magically get the lock or lose
it, all transitions are explicit.