Add an initial disk to the semantics
The initial disk is needed since following the protocol within a locked
region is a bit nuanced: we won't guarantee the invariant for the
current disk (it will be for an older disk), but this doesn't matter
since concurrent threads can't actually observe any changes to the disk
until we release the unlock. Since other threads are insensitive in some
sense we should be able to use any (intermediate) disk.

These semantics don't have quite the right obligations in terms of which
disk is used for Assgn and Unlock. In practice the code has one case of
using this:

lock; run optimistic p; start read fill; update cache; unlock

The cache update should go from the post-lock disk to the current disk,
while the unlock just exposes the current disk and respects the global
invariant. It's hard to justify these as obligations though (why update
the current disk on assignment? it's still not exposed...), and there's
no consistency proof to check us. It seems like we could even decide
whether or not we want to update the initial disk at each assignment in
the code, as a ghost operation.