optimize writing out log descriptor blocks
the log descriptor blocks contain 512 64-bit addresses, but most are
zero (when flushing a small set of changes).  we used to flush the
block by padding it out with zero elements, and pickling all of them.

this change removes the padding; the code already does the right thing
by filling in zeroes when pickling an empty list.  no proofs yet.

improves in-memory mailbench from 330 msg/s to 367 msg/s.