Ass unlocked invariant that d_i = d
In the process cleaned up fs_invariant with a notation to capture common
case of applying to parts of a sigma (though we want to use the unfolded
form for the most part, so it's a notation and not a definition).

The startLocked helper now goes through with its (mostly untested) spec.
We might need to give unlock a custom spec to make finishing the proof
tractable.