Expand on write_syscall_ok proof a bit
This is a fairly complicated proof due to every inariant obligation
being proven directly, as well as from having a fair number of cases and
control flow paths. Need to factor out a spec (Assgn1 cache; Unlock is a
good candidate) and perhaps some special-case theorems about the
guarantee. Might also help for Assgn to copy over the guarantee from the
precondition hypothesis that guarantees it, to avoid duplicate proof
effort.