Start proving file_set_attr
Need alter_inum function and equivalence to update_subtree (from
io-concur branch).

The requirement that the update used for the ghost update match the spec
directory update is too strong: it should be allowed to assume the
precondition and a homedir_rely step.

The requirement that the update follow the protocol is only satisfied if
the home directories are disjoint, since otherwise this update might be
disallowed by some other overlapping home directory.