Prove top-level file_set_attr_ok
Need a new notion of home directory disjointness, defined as a given
home directory being disjoint from all others. Directory disjointness
itself is defined in terms of a diverging location in the paths, which
is particularly convenient to work with in proofs. Without this
disjointness condition the home directories still work, but there are no
allowed updates to any overlapping home directories (each precludes the
other).

This completes an admit-free proof of the file_set_attr top-level spec,
albeit only if the current thread's home directory is appropriately
isolated. The assumption that the home directory is isolated is left in
the precondition rather than being a property of all the home
directories to avoid defining an infinite set of disjoint home
directories, even though not all threads will be used (though not
impossible this is annoying).