Change BmapAllocCache rep and prove operations
Like for the DirCache per-directory caches, this uses an optional cache;
if cached data is present, it is assumed correct. Also, an empty and
present cache is now useful; in the previous iteration, an empty cache
meant 'not present, go and fetch'.

The rep is a little convoluted; the problem is that the bottom-level
allocator's rep doesn't special case block 0, even though the actual
allocator function will never return it (see the postcondition for
alloc). Changing the rep would have required an annoying amount of
changes above the block allocator, so instead, the cache rep states that
a present cache is identical to the on-disk state except for block 0,
which is ignored if present in the on-disk state and which can never
appear in the cache.